Add files via upload
for the DR4-II Nose piece, there are two versions.  The first version that I made, the opening to take the GPS was a little bit small.  I used a dremmel and just opened it up some and it worked fine.  I did try to make another version, with a larger opening, which is just made as big as I possibly could, and the GPS fits nicer but it is obviously a give and take situation.  I personally like the first version better with the post print dremmeling to make it fit.  I will provide both and leave the decision up to you.   The second version is the one with "alpha 0.2" in the title.