Add files via upload
This 'gimbal adapter plate' needs to be strengthened on the arm.  I took a piece of scrap plastic about 2mm thick and used a heat gun to bend it to match the shape.  Then I glued it with CA and also melted the edges together so that it loos like one piece.  It is incredibly strong after this, much stronger I believe than if it were printed this way.  If anyone wants detail pics, you know where to find me.  Send me a message.